src/text_overlay/requirements.txt
albumentations
nltk
numpy>=2.0.0
pillow>=10.3.0
rake-nltk
torch>=2.3.1
tqdm>=4.66.4
